Move reference mapping functions to common
No change in performance. This allows the decoder to compute
the mapping to avoid signalling it explicitly in the bitstream.
