+++ av1/encoder/encode_strategy.c
@@ -1005,243 +1005,6 @@ static int denoise_and_encode(AV1_COMP *const cpi, uint8_t *const dest,
   return AOM_CODEC_OK;
 }
 
-/*!\cond */
-// Struct to keep track of relevant reference frame data
-typedef struct {
-  int map_idx;
-  int disp_order;
-  int pyr_level;
-  int used;
-} RefBufMapData;
-/*!\endcond */
-
-// Comparison function to sort reference frames in ascending display order
-static int compare_map_idx_pair_asc(const void *a, const void *b) {
-  if (((RefBufMapData *)a)->disp_order == ((RefBufMapData *)b)->disp_order) {
-    return 0;
-  } else if (((const RefBufMapData *)a)->disp_order >
-             ((const RefBufMapData *)b)->disp_order) {
-    return 1;
-  } else {
-    return -1;
-  }
-}
-
-// Checks to see if a particular reference frame is already in the reference
-// frame map
-static int is_in_ref_map(RefBufMapData *map, int disp_order, int n_frames) {
-  for (int i = 0; i < n_frames; i++) {
-    if (disp_order == map[i].disp_order) return 1;
-  }
-  return 0;
-}
-
-// Add a reference buffer index to a named reference slot
-static void add_ref_to_slot(RefBufMapData *ref, int *const remapped_ref_idx,
-                            int frame) {
-  remapped_ref_idx[frame - LAST_FRAME] = ref->map_idx;
-  ref->used = 1;
-}
-
-// Threshold dictating when we are allowed to start considering
-// leaving lowest level frames unmapped
-#define LOW_LEVEL_FRAMES_TR 5
-
-// Find which reference buffer should be left out of the named mapping.
-// This is because there are 8 reference buffers and only 7 named slots.
-static void set_unmapped_ref(RefBufMapData *buffer_map, int n_bufs,
-                             int n_min_level_refs, int min_level,
-                             int cur_frame_disp) {
-  int max_dist = 0;
-  int unmapped_idx = -1;
-  if (n_bufs <= ALTREF_FRAME) return;
-  for (int i = 0; i < n_bufs; i++) {
-    if (buffer_map[i].used) continue;
-    if (buffer_map[i].pyr_level != min_level ||
-        n_min_level_refs >= LOW_LEVEL_FRAMES_TR) {
-      int dist = abs(cur_frame_disp - buffer_map[i].disp_order);
-      if (dist > max_dist) {
-        max_dist = dist;
-        unmapped_idx = i;
-      }
-    }
-  }
-  assert(unmapped_idx >= 0 && "Unmapped reference not found");
-  buffer_map[unmapped_idx].used = 1;
-}
-
-void av1_get_ref_frames(AV1_COMP *const cpi, int cur_frame_disp,
-                        RefFrameMapPair ref_frame_map_pairs[REF_FRAMES]) {
-  AV1_COMMON *cm = &cpi->common;
-  int *const remapped_ref_idx = cm->remapped_ref_idx;
-
-  int buf_map_idx = 0;
-
-  // Initialize reference frame mappings
-  for (int i = 0; i < REF_FRAMES; ++i) remapped_ref_idx[i] = INVALID_IDX;
-
-  RefBufMapData buffer_map[REF_FRAMES];
-  int n_bufs = 0;
-  memset(buffer_map, 0, REF_FRAMES * sizeof(buffer_map[0]));
-  int min_level = MAX_ARF_LAYERS;
-  int max_level = 0;
-
-  // Go through current reference buffers and store display order, pyr level,
-  // and map index.
-  for (int map_idx = 0; map_idx < REF_FRAMES; map_idx++) {
-    // Get reference frame buffer
-    RefFrameMapPair ref_pair = ref_frame_map_pairs[map_idx];
-    if (ref_pair.disp_order == -1) continue;
-    const int frame_order = ref_pair.disp_order;
-    // Avoid duplicates
-    if (is_in_ref_map(buffer_map, frame_order, n_bufs)) continue;
-    const int reference_frame_level = ref_pair.pyr_level;
-
-    // Keep track of the lowest and highest levels that currently exist
-    if (reference_frame_level < min_level) min_level = reference_frame_level;
-    if (reference_frame_level > max_level) max_level = reference_frame_level;
-
-    buffer_map[n_bufs].map_idx = map_idx;
-    buffer_map[n_bufs].disp_order = frame_order;
-    buffer_map[n_bufs].pyr_level = reference_frame_level;
-    buffer_map[n_bufs].used = 0;
-    n_bufs++;
-  }
-
-  // Sort frames in ascending display order
-  qsort(buffer_map, n_bufs, sizeof(buffer_map[0]), compare_map_idx_pair_asc);
-
-  int n_min_level_refs = 0;
-  int n_past_high_level = 0;
-  int closest_past_ref = -1;
-  int golden_idx = -1;
-  int altref_idx = -1;
-
-  // Find the GOLDEN_FRAME and BWDREF_FRAME.
-  // Also collect various stats about the reference frames for the remaining
-  // mappings
-  for (int i = n_bufs - 1; i >= 0; i--) {
-    if (buffer_map[i].pyr_level == min_level) {
-      // Keep track of the number of lowest level frames
-      n_min_level_refs++;
-      if (buffer_map[i].disp_order < cur_frame_disp && golden_idx == -1 &&
-          remapped_ref_idx[GOLDEN_FRAME - LAST_FRAME] == INVALID_IDX) {
-        // Save index for GOLDEN
-        golden_idx = i;
-      } else if (buffer_map[i].disp_order > cur_frame_disp &&
-                 altref_idx == -1 &&
-                 remapped_ref_idx[ALTREF_FRAME - LAST_FRAME] == INVALID_IDX) {
-        // Save index for ALTREF
-        altref_idx = i;
-      }
-    } else if (buffer_map[i].disp_order == cur_frame_disp) {
-      // Map the BWDREF_FRAME if this is the show_existing_frame
-      add_ref_to_slot(&buffer_map[i], remapped_ref_idx, BWDREF_FRAME);
-    }
-
-    // Keep track of the number of past frames that are not at the lowest level
-    if (buffer_map[i].disp_order < cur_frame_disp &&
-        buffer_map[i].pyr_level != min_level)
-      n_past_high_level++;
-
-    // Keep track of where the frames change from being past frames to future
-    // frames
-    if (buffer_map[i].disp_order < cur_frame_disp && closest_past_ref < 0)
-      closest_past_ref = i;
-  }
-
-  // Do not map GOLDEN and ALTREF based on their pyramid level if all reference
-  // frames have the same level
-  if (n_min_level_refs <= n_bufs) {
-    // Map the GOLDEN_FRAME
-    if (golden_idx > -1)
-      add_ref_to_slot(&buffer_map[golden_idx], remapped_ref_idx, GOLDEN_FRAME);
-    // Map the ALTREF_FRAME
-    if (altref_idx > -1)
-      add_ref_to_slot(&buffer_map[altref_idx], remapped_ref_idx, ALTREF_FRAME);
-  }
-
-  // Find the buffer to be excluded from the mapping
-  set_unmapped_ref(buffer_map, n_bufs, n_min_level_refs, min_level,
-                   cur_frame_disp);
-
-  // Place past frames in LAST_FRAME, LAST2_FRAME, and LAST3_FRAME
-  for (int frame = LAST_FRAME; frame < GOLDEN_FRAME; frame++) {
-    // Continue if the current ref slot is already full
-    if (remapped_ref_idx[frame - LAST_FRAME] != INVALID_IDX) continue;
-    // Find the next unmapped reference buffer
-    // in decreasing ouptut oreder relative to current picture
-    int next_buf_max = 0;
-    int next_disp_order = INT_MIN;
-    for (buf_map_idx = n_bufs - 1; buf_map_idx >= 0; buf_map_idx--) {
-      if (!buffer_map[buf_map_idx].used &&
-          buffer_map[buf_map_idx].disp_order < cur_frame_disp &&
-          buffer_map[buf_map_idx].disp_order > next_disp_order) {
-        next_disp_order = buffer_map[buf_map_idx].disp_order;
-        next_buf_max = buf_map_idx;
-      }
-    }
-    buf_map_idx = next_buf_max;
-    if (buf_map_idx < 0) break;
-    if (buffer_map[buf_map_idx].used) break;
-    add_ref_to_slot(&buffer_map[buf_map_idx], remapped_ref_idx, frame);
-  }
-
-  // Place future frames (if there are any) in BWDREF_FRAME and ALTREF2_FRAME
-  for (int frame = BWDREF_FRAME; frame < REF_FRAMES; frame++) {
-    // Continue if the current ref slot is already full
-    if (remapped_ref_idx[frame - LAST_FRAME] != INVALID_IDX) continue;
-    // Find the next unmapped reference buffer
-    // in increasing ouptut oreder relative to current picture
-    int next_buf_max = 0;
-    int next_disp_order = INT_MAX;
-    for (buf_map_idx = n_bufs - 1; buf_map_idx >= 0; buf_map_idx--) {
-      if (!buffer_map[buf_map_idx].used &&
-          buffer_map[buf_map_idx].disp_order > cur_frame_disp &&
-          buffer_map[buf_map_idx].disp_order < next_disp_order) {
-        next_disp_order = buffer_map[buf_map_idx].disp_order;
-        next_buf_max = buf_map_idx;
-      }
-    }
-    buf_map_idx = next_buf_max;
-    if (buf_map_idx < 0) break;
-    if (buffer_map[buf_map_idx].used) break;
-    add_ref_to_slot(&buffer_map[buf_map_idx], remapped_ref_idx, frame);
-  }
-
-  // Place remaining past frames
-  buf_map_idx = closest_past_ref;
-  for (int frame = LAST_FRAME; frame < REF_FRAMES; frame++) {
-    // Continue if the current ref slot is already full
-    if (remapped_ref_idx[frame - LAST_FRAME] != INVALID_IDX) continue;
-    // Find the next unmapped reference buffer
-    for (; buf_map_idx >= 0; buf_map_idx--) {
-      if (!buffer_map[buf_map_idx].used) break;
-    }
-    if (buf_map_idx < 0) break;
-    if (buffer_map[buf_map_idx].used) break;
-    add_ref_to_slot(&buffer_map[buf_map_idx], remapped_ref_idx, frame);
-  }
-
-  // Place remaining future frames
-  buf_map_idx = n_bufs - 1;
-  for (int frame = ALTREF_FRAME; frame >= LAST_FRAME; frame--) {
-    // Continue if the current ref slot is already full
-    if (remapped_ref_idx[frame - LAST_FRAME] != INVALID_IDX) continue;
-    // Find the next unmapped reference buffer
-    for (; buf_map_idx > closest_past_ref; buf_map_idx--) {
-      if (!buffer_map[buf_map_idx].used) break;
-    }
-    if (buf_map_idx < 0) break;
-    if (buffer_map[buf_map_idx].used) break;
-    add_ref_to_slot(&buffer_map[buf_map_idx], remapped_ref_idx, frame);
-  }
-
-  // Fill any slots that are empty (should only happen for the first 7 frames)
-  for (int i = 0; i < REF_FRAMES; ++i)
-    if (remapped_ref_idx[i] == INVALID_IDX) remapped_ref_idx[i] = 0;
-}
-
 int av1_encode_strategy(AV1_COMP *const cpi, size_t *const size,
                         uint8_t *const dest, unsigned int *frame_flags,
                         int64_t *const time_stamp, int64_t *const time_end,
@@ -1418,7 +1181,8 @@ int av1_encode_strategy(AV1_COMP *const cpi, size_t *const size,
                                force_refresh_all);
 
   RefFrameMapPair ref_frame_map_pairs[REF_FRAMES];
-  init_ref_map_pair(cpi, ref_frame_map_pairs);
+  init_ref_map_pair(&cpi->common, ref_frame_map_pairs,
+                    gf_group->update_type[gf_group->index] == KF_UPDATE);
 
   if (!is_stat_generation_stage(cpi)) {
     const RefCntBuffer *ref_frames[INTER_REFS_PER_FRAME];
@@ -1428,7 +1192,7 @@ int av1_encode_strategy(AV1_COMP *const cpi, size_t *const size,
       const int order_offset = gf_group->arf_src_offset[gf_group->index];
       const int cur_frame_disp =
           cpi->common.current_frame.frame_number + order_offset;
-      av1_get_ref_frames(cpi, cur_frame_disp, ref_frame_map_pairs);
+      av1_get_ref_frames(cm, cur_frame_disp, ref_frame_map_pairs);
     }
 
     // Get the reference frames
